Add the `alloc::prelude` module
It contains the re-exports that are in `std::prelude::v1`
but not in `core::prelude::v1`.

Calling it prelude is somewhat of a misnomer since (unlike those modules
in `std` or `core`) its contents are never implicitly imported in modules.
Rather it is intended to be used with an explicit glob import like
`use alloc::prelude::*;`.
However there is precedent for the same misnomer with `std::io::prelude`,
for example.

This new module is insta-stable, but only exists in a crate that is,
at this point, unstable. RFC https://github.com/rust-lang/rfcs/pull/2480
proposes stabilizing the crate (with this new `prelude` module).